requirements.txt
tensorflow
torch
pandas
numpy
streamlit
scikit-learn
matplotlib
notebook
langchain
langchain_community
langchain-google-genai
langchain-ollama
langchain-groq
sentence-transformers
transformers
huggingface_hub
pymupdf
faiss-cpu
python-dotenv
langgraph
duckduckgo-search
wikipedia
langchainhub
ddgs